% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_biomass_sok}
\alias{calc_biomass_sok}
\title{Calculate spawning biomass from spawn-on-kelp (SOK) harvest.}
\usage{
calc_biomass_sok(
  SOK,
  nu = pars$SOK$nu,
  upsilon = pars$SOK$upsilon,
  M = pars$SOK$M,
  theta = calc_egg_conversion()
)
}
\arguments{
\item{SOK}{Numeric. Weight of spawn-on-kelp (SOK) harvest in kilograms.}

\item{nu}{Numeric. Proportion of SOK product that is kelp; from
\code{\link{pars}}.}

\item{upsilon}{Numeric. SOK product weight increase due to brining as a
proportion; from \code{\link{pars}}.}

\item{M}{Numeric. Average weight in kilograms of a fertilized egg; from
\code{\link{pars}}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{calc_egg_conversion}}.}
}
\value{
Numeric. Spawning biomass in tonnes.
}
\description{
Calculate spawning biomass in tonnes from spawn-on-kelp (SOK) harvest in
kilograms.
}
\examples{
data(pars)
calc_biomass_sok(SOK = 100)
}
\seealso{
\code{\link{calc_egg_conversion}} \code{\link{pars}}
}
