% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_egg_conversion}
\alias{calc_egg_conversion}
\title{Convert number of eggs to spawning biomass.}
\usage{
calc_egg_conversion(
  omega = pars$conversion$omega,
  female = pars$conversion$female
)
}
\arguments{
\item{omega}{Numeric. The number of eggs per kilogram of female spawners;
from \code{\link{pars}}.}

\item{female}{Numeric. The proportion of spawners that are female; from
\code{\link{pars}}.}
}
\value{
Numeric. The conversion factor for eggs to spawn index in tonnes
  (i.e., biomass). Divide the number of eggs by the conversion factor to get
  biomass.
}
\description{
Calculate the conversion factor for the number of Pacific Herring eggs to the
spawn index (i.e., biomass) in tonnes.
}
\examples{
data(pars)
calc_egg_conversion()
}
\seealso{
\code{\link{pars}}
}
