% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_macro_index}
\alias{calc_macro_index}
\title{Calculate the Macrocystis spawn index.}
\usage{
calc_macro_index(
  where,
  areas,
  years,
  chi = 2,
  xi = pars$macrocystis$xi,
  gamma = pars$macrocystis$gamma,
  delta = pars$macrocystis$delta,
  epsilon = pars$macrocystis$epsilon,
  theta = eggs_to_sb(),
  quiet = FALSE
)
}
\arguments{
\item{where}{List. Location and name of the Pacific Herring spawn survey
database, as well as required tables (see examples).}

\item{areas}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{load_area_data}}.}

\item{years}{Numeric vector. Years(s) to include in the calculations. Message
if < \code{pars$years$assess}.}

\item{chi}{Numeric. Transect swath (i.e., width) in metres.}

\item{xi}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{gamma}{Numeric. Regression exponent on egg layers; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{delta}{Numeric. Regression exponent on plant height; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{epsilon}{Numeric. Regression exponent on number of stalks per plant;
from \code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{eggs_to_sb}}. Message if < 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
List. The element \code{si} is a tibble with Macrocystis spawn index
(\code{MacroSI}) in tonnes by spawn number and year. The spawn number is
the finest spatial scale at which we calculate the spawn index. Other
information in this tibble comes from \code{a}: Region, Statistical Area,
Section, and Location code.
}
\description{
Calculate the Pacific Herring Macrocystis spawn index in tonnes
\insertCite{HaegeleSchweigert1990}{SpawnIndex}. This function primarily
wrangles and prepares the data; the actual calculation is done by
\code{\link{eggs_macro}}.
}
\note{
The `spawn index' is a relative index of spawning biomass.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
data(pars)
macro_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(
    all_spawn = "tSSAllspawn", plants = "tSSMacPlant",
    transects = "tSSMacTrans"
  )
)
macro_spawn <- calc_macro_index(
  where = macro_loc, areas = areas, years = 2010:2015
)
macro_spawn$si
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{load_area_data}}
\code{\link{eggs_to_sb}} \code{\link{pars}}
\code{\link{eggs_macro}}

Other calculation functions: 
\code{\link{calc_sok_sb}()},
\code{\link{calc_surf_index}()},
\code{\link{calc_under_index}()},
\code{\link{dens_surf}()},
\code{\link{dens_under_alg}()},
\code{\link{dens_under_sub}()},
\code{\link{eggs_macro}()},
\code{\link{eggs_to_sb}()}
}
\concept{calculation functions}
