% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_macro_spawn}
\alias{calc_macro_spawn}
\title{Calculate the Macrocystis spawn index.}
\usage{
calc_macro_spawn(
  where,
  a,
  yrs,
  t_swath = 2,
  xi = pars$macrocystis$xi,
  gamma = pars$macrocystis$gamma,
  delta = pars$macrocystis$delta,
  epsilon = pars$macrocystis$epsilon,
  theta = calc_egg_conversion(),
  quiet = FALSE
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring Macrocystis spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{load_area_data}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in \code{pars$years$assess}.}

\item{t_swath}{Numeric. Transect swath (i.e., width) in metres.}

\item{xi}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{gamma}{Numeric. Regression exponent on egg layers; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{delta}{Numeric. Regression exponent on plant height; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{epsilon}{Numeric. Regression exponent on number of stalks per plant;
from \code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{calc_egg_conversion}}.}

\item{quiet}{Logical. Set to TRUE to prevent messages; default is FALSE.}
}
\value{
List. The element \code{SI} is a tibble with Macrocystis spawn index
(\code{MacroSI}) in tonnes by spawn number and year. The spawn number is
the finest spatial scale at which we calculate the spawn index. Other
information in this tibble comes from \code{a}: Region, Statistical Area,
Section, and Location code.
}
\description{
Calculate the Pacific Herring Macrocystis spawn index in tonnes.
}
\note{
The `spawn index' is a relative index of spawning biomass.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
data(pars)
macro_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(
    all_spawn = "tSSAllspawn", plants = "tSSMacPlant",
    transects = "tSSMacTrans"
  )
)
macro_spawn <- calc_macro_spawn(
  where = macro_loc, a = areas, yrs = 2010:2015
)
macro_spawn$SI
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{load_area_data}}
\code{\link{calc_egg_conversion}} \code{\link{pars}}

Other calculation functions: 
\code{\link{calc_biomass_sok}()},
\code{\link{calc_egg_conversion}()},
\code{\link{calc_surf_spawn}()},
\code{\link{calc_under_spawn}()}
}
\concept{calculation functions}
