% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_sok_index}
\alias{calc_sok_index}
\title{Calculate the spawn index from spawn-on-kelp (SOK) harvest.}
\usage{
calc_sok_index(
  sok,
  nu = pars$sok$nu,
  upsilon = pars$sok$upsilon,
  egg_weight = pars$sok$egg_weight,
  theta = eggs_to_sb(),
  quiet = FALSE
)
}
\arguments{
\item{sok}{Numeric. Weight of spawn-on-kelp (SOK) harvest in kilograms.
Message if < 0.}

\item{nu}{Numeric. Proportion of SOK product that is kelp; from
\code{\link{pars}}. Message if < 0 and/or > 1.}

\item{upsilon}{Numeric. SOK product weight increase due to brining as a
proportion; from \code{\link{pars}}. Message if < 0 and/or > 1.}

\item{egg_weight}{Numeric. Average weight in kilograms of a fertilized egg;
from \code{\link{pars}}. Message if < 0.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{eggs_to_sb}}. Message if < 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. Spawn index in tonnes. Message if < 0.
}
\description{
Calculate the spawn index in tonnes from spawn-on-kelp (SOK) harvest in
kilograms.
}
\examples{
data(pars)
calc_sok_index(sok = 100)
}
\seealso{
\code{\link{eggs_to_sb}} \code{\link{pars}}

Other calculation functions: 
\code{\link{calc_macro_index}()},
\code{\link{calc_surf_index}()},
\code{\link{calc_under_index}()},
\code{\link{dens_surf}()},
\code{\link{dens_under_alg}()},
\code{\link{dens_under_sub}()},
\code{\link{eggs_macro}()},
\code{\link{eggs_to_sb}()}
}
\concept{calculation functions}
