% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_surf_spawn}
\alias{calc_surf_spawn}
\title{Calculate the surface spawn index.}
\usage{
calc_surf_spawn(
  where,
  a,
  widths,
  yrs,
  intense = intensity,
  intense_yrs = pars$years$survey:pars$years$layers,
  rescale_yrs = pars$years$assess:pars$years$assess,
  alpha = pars$surface$alpha,
  beta = pars$surface$beta,
  theta = calc_egg_conversion(),
  quiet = FALSE
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring surface spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{load_area_data}}.}

\item{widths}{List. List of three tables: median region, section, and pool
widths in metres (m); from \code{\link{get_width}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations. Message
if 1951s`.}

\item{intense}{Tibble. Table of spawn intensity categories and number of egg
layers; from \code{\link{intensity}}.}

\item{intense_yrs}{Numeric vector. Years where intensity categories are used
to determine egg layers. Message if >1979s`.}

\item{rescale_yrs}{Numeric vector. Years where intensity needs to be
re-scaled from 5 to 9 categories. Message if >1951s`.}

\item{alpha}{Numeric. Regression intercept; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{beta}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{calc_egg_conversion}}.}

\item{quiet}{Logical. Set to TRUE to prevent messages; default is FALSE.}
}
\value{
List. The element \code{SI} is a tibble with surface spawn index
(\code{SurfSI}) in tonnes by spawn number and year. The spawn number is the
finest spatial scale at which we calculate the spawn index. Other
information in this tibble comes from \code{a}: Region, Statistical Area,
Section, and Location code.
}
\description{
Calculate the Pacific Herring surface spawn index in tonnes.
}
\note{
The `spawn index' is a relative index of spawning biomass.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
width_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(
    region_std = "RegionStd", section_std = "SectionStd",
    pool_std = "PoolStd"
  )
)
width_bar <- get_width(where = width_loc, a = areas)
data(pars)
data(intensity)
surf_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(surface = "tSSSurface", all_spawn = "tSSAllspawn")
)
surf_spawn <- calc_surf_spawn(
  where = surf_loc, a = areas, widths = width_bar, yrs = 2010:2015
)
surf_spawn$SI
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{load_area_data}}
\code{\link{get_width}} \code{\link{calc_egg_conversion}}
\code{\link{pars}} \code{\link{intensity}}

Other calculation functions: 
\code{\link{calc_biomass_sok}()},
\code{\link{calc_egg_conversion}()},
\code{\link{calc_macro_spawn}()},
\code{\link{calc_under_spawn}()}
}
\concept{calculation functions}
