% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_under_index}
\alias{calc_under_index}
\title{Calculate the understory spawn index.}
\usage{
calc_under_index(
  where,
  areas,
  years,
  alg_coefs = algae_coefs,
  tau = under_width_facs,
  varphi = pars$understory$varphi,
  vartheta = pars$understory$vartheta,
  varrho = pars$under$varrho,
  varsigma = pars$understory$varsigma,
  theta = eggs_to_sb(),
  quiet = FALSE
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring understory spawn database
(see examples).}

\item{areas}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{load_area_data}}.}

\item{years}{Numeric vector. Years(s) to include in the calculations. Message
if < \code{pars$years$assess}.}

\item{alg_coefs}{Tibble. Table of algae coefficients; from
\code{\link{algae_coefs}}.}

\item{tau}{Tibble. Table of understory spawn width adjustment factors from
\code{\link{under_width_facs}}.}

\item{varphi}{Numeric. Regression slope for substrate; from
\code{\link{pars}} \insertCite{HaegeleEtal1979}{SpawnIndex}.}

\item{vartheta}{Numeric. Regression slope for algae; from \code{\link{pars}}
\insertCite{Schweigert2005}{SpawnIndex}.}

\item{varrho}{Numeric. Regression exponent on number of egg layers; from
\code{\link{pars}} \insertCite{Schweigert2005}{SpawnIndex}.}

\item{varsigma}{Numeric. Regression exponent on proportion of algae; from
\code{\link{pars}} \insertCite{Schweigert2005}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{eggs_to_sb}}. Message if < 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
List. The element \code{si} is a tibble with understory spawn index
(\code{UnderSI}) in tonnes by spawn number and year. The spawn number is
the finest spatial scale at which we calculate the spawn index. Other
information in this tibble comes from \code{a}: Region, Statistical Area,
Section, and Location code.
}
\description{
Calculate the Pacific Herring understory spawn index in tonnes
\insertCite{HaegeleEtal1979,Schweigert2005}{SpawnIndex}. This function
primarily wrangles and prepares the data; the actual calculations are done by
\code{\link{dens_under_sub}} and \code{\link{dens_under_alg}}.
}
\note{
The `spawn index' is a relative index of spawning biomass.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
under_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(
    all_spawn = "tSSAllspawn", alg_trans = "tSSVegTrans",
    stations = "tSSStations", algae = "tSSVegetation"
  )
)
data(under_width_facs)
data(pars)
data(algae_coefs)
under_spawn <- calc_under_index(
  where = under_loc, areas = areas, years = 2010:2015
)
under_spawn$si
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{load_area_data}}
\code{\link{eggs_to_sb}} \code{\link{pars}}
\code{\link{algae_coefs}} \code{\link{dens_under_sub}}
\code{\link{dens_under_alg}}

Other calculation functions: 
\code{\link{calc_macro_index}()},
\code{\link{calc_sok_sb}()},
\code{\link{calc_surf_index}()},
\code{\link{dens_surf}()},
\code{\link{dens_under_alg}()},
\code{\link{dens_under_sub}()},
\code{\link{eggs_macro}()},
\code{\link{eggs_to_sb}()}
}
\concept{calculation functions}
