% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcEggConversion}
\alias{CalcEggConversion}
\title{Convert number of eggs to spawning biomass.}
\usage{
CalcEggConversion(fecundity = 2e+05, pFemale = 0.5)
}
\arguments{
\item{fecundity}{Numeric. The number of eggs per kilogram of female spawners
\insertCite{Hay1985,HayBrett1988}{SpawnIndex}.}

\item{pFemale}{Numeric. The proportion of spawners that are female.}
}
\value{
Numeric. The conversion factor for eggs to spawn index in tonnes
  (i.e., biomass). Divide the number of eggs by the conversion factor to get
  biomass.
}
\description{
Calculate the conversion factor for Pacific Herring, to convert the number of
eggs to the spawn index (i.e., biomass) in tonnes.
}
\examples{
CalcEggConversion()
}
\references{
\insertAllCited{}
}
