% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcMacroSpawn}
\alias{CalcMacroSpawn}
\title{Calculate the Macrocystis spawn index.}
\usage{
CalcMacroSpawn(
  where,
  a,
  yrs,
  tSwath = 2,
  beta = pars$macrocystis$beta,
  gamma = pars$macrocystis$gamma,
  delta = pars$macrocystis$delta,
  epsilon = pars$macrocystis$epsilon,
  theta = CalcEggConversion()
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring Macrocystis spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations; from
\code{\link{LoadAreaData}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{tSwath}{Numeric. Transect swath (i.e., width) in metres.}

\item{beta}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{gamma}{Numeric. Regression exponent on egg layers; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{delta}{Numeric. Regression exponent on plant height; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{epsilon}{Numeric. Regression exponent on numnber of stalks per plant;
from \code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{CalcEggConversion}}.}
}
\value{
List. The element \code{SI} is a tibble with Macrocystis spawn index
  (\code{MacroSI}) in tonnes by spawn number and year. The spawn number is
  the finest spatial scale at which we calculate the spawn index. Other
  information in this tibble comes from \code{a}: Region, Statistical Area,
  Section, and Location code.
}
\description{
Calculate the Pacific Herring Macrocystis spawn index in tonnes.
}
\note{
The `spawn index' is a relative index of spawning biomass.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
data(pars)
macroLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(
    allSpawn = "tSSAllspawn", plants = "tSSMacPlant",
    transects = "tSSMacTrans"
  )
)
macroSpawn <- CalcMacroSpawn(where = macroLoc, a = areas, yrs = 2010:2015)
macroSpawn$SI
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{LoadAreaData}}
  \code{\link{CalcEggConversion}} \code{\link{pars}}
}
