% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcSurfSpawn}
\alias{CalcSurfSpawn}
\title{Calculate the surface spawn index.}
\usage{
CalcSurfSpawn(
  where,
  a,
  yrs,
  intYrs = yrs[yrs < 1979],
  rsYrs = intYrs[intYrs < 1951],
  f
)
}
\arguments{
\item{where}{Location of the Pacific Herring surface spawn database.}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations. Returned from
\code{\link{LoadAreaData}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{intYrs}{Numeric vector. Years where intensity categores are used to
determine egg layers (default yrs[yrs < 1979]).}

\item{rsYrs}{Numeric vector. Years where intensity needs to be re-scaled
(default intYrs[intYrs < 1951]).}

\item{f}{Egg conversion factor (eggs to biomass) from
\code{\link{CalcEggConversion}}.}
}
\value{
Numeric. Surface spawn index in tonnes.
}
\description{
Calculate the Pacific Herring surface spawn index in tonnes. Note that the
'spawn index' is a relative index of spawning biomass.
}
\examples{
areaLoc <- list( loc="data", db="HerringSpawn.mdb",
  fns=list(sections="Sections", locations="Location") )
inCRS <- "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
outCRS <- "+proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000
  +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
areas <- LoadAreaData( reg="WCVI", sectionSub=NA, where=areaLoc, inCRS=inCRS,
                       outCRS=outCRS )
surfLoc <- list( loc="data", db="HerringSpawn.mdb",
  fns=list(regionStd="RegionStd", sectionStd="SectionStd", poolStd="PoolStd",
           surface="tSSSurface", intensity="Intensity",
           allSpawn="tSSAllspawn") )
yrRange <- 2010:2015  # Year range to include
ECF <- CalcEggConversion( )
CalcSurfSpawn( where=surfLoc, a=areas, yrs=yrRange, f=ECF )
}
\references{
\insertRef{SchweigertEtal1997}{SpawnIndex}
}
\seealso{
\code{\link{LoadAreaData}} \code{\link{CalcEggConversion}}
}
