% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_tibble}
\alias{check_tibble}
\title{Check that an object is a tibble.}
\usage{
check_tibble(dat, min_rows = 1, quiet = FALSE)
}
\arguments{
\item{dat}{List. Message if no rows; stop if not a tibble.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
NULL.
}
\description{
Check that an object is a tibble and has a minimum number of rows. Gives a
message if there are fewer rows, and an error if it's not a tibble.
}
\examples{
alpha <- data.frame(a = 1, b = 2)
beta <- tibble()
check_tibble(dat = list(alpha = alpha, beta = beta))
}
\seealso{
Other helper functions: 
\code{\link{check_numeric}()}
}
\concept{helper functions}
