% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_where}
\alias{check_where}
\title{Check a `where' object.}
\usage{
check_where(dat, dat_names)
}
\arguments{
\item{dat}{List. Where object: list of characters with names (see examples).}

\item{dat_names}{Character vector. Names of list items.}
}
\value{
NULL.
}
\description{
Check that a `where' object is a list of characters with required names.
Error if this is not the case.
}
\examples{
check_where(dat = list(x = "one", y = "two"), dat_names = c("x", "y"))
}
\seealso{
Other check functions: 
\code{\link{check_numeric}()},
\code{\link{check_tibble}()}
}
\concept{check functions}
