% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{dens_surf}
\alias{dens_surf}
\title{Calculate surface spawn egg density.}
\usage{
dens_surf(
  alpha = pars$surface$alpha,
  beta = pars$surface$beta,
  egg_layers,
  quiet = FALSE
)
}
\arguments{
\item{alpha}{Numeric. Regression intercept; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{beta}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{egg_layers}{Numeric. Number of egg layers. Message if < 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. Egg density in thousands of eggs per square metre. Message
if <= 0.
}
\description{
Calculate Pacific Herring surface spawn egg density in thousands of eggs per
square metre (10^3 * eggs / m^2). This function implements a linear
regression model to estimate egg density; implemented in
\code{\link{calc_surf_index}} \insertCite{SchweigertEtal1997}{SpawnIndex}.
}
\note{
Density is set to 0 when egg layers is 0 (otherwise it would be
\code{alpha}). Also, there is an error in
\insertCite{SchweigertEtal1997;textual}{SpawnIndex}; surface egg density is
in thousands of eggs per square metre.
}
\examples{
data(pars)
dens_surf(egg_layers = 4)
}
\references{
\insertAllCited
}
\seealso{
\code{\link{pars}} \code{\link{calc_surf_index}}

Other calculation functions: 
\code{\link{calc_macro_index}()},
\code{\link{calc_sok_sb}()},
\code{\link{calc_surf_index}()},
\code{\link{calc_under_index}()},
\code{\link{dens_under_alg}()},
\code{\link{dens_under_sub}()},
\code{\link{eggs_macro}()},
\code{\link{eggs_to_sb}()}
}
\concept{calculation functions}
