% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{dens_under_sub}
\alias{dens_under_sub}
\title{Calculate understory spawn egg density on substrate.}
\usage{
dens_under_sub(
  varphi = pars$understory$varphi,
  sub_layers,
  sub_prop,
  quiet = FALSE
)
}
\arguments{
\item{varphi}{Numeric. Regression slope for substrate; from
\code{\link{pars}} \insertCite{HaegeleEtal1979}{SpawnIndex}.}

\item{sub_layers}{Numeric. Number of egg layers on substrate. Message if < 0.}

\item{sub_prop}{Numeric. Proportion of substrate covered in eggs. Message if
< 0 or > 1.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. Egg density in thousands of eggs per square metre. Message
if < 0.
}
\description{
Calculate Pacific Herring understory spawn egg density on substrate in
thousands of eggs per square metre (10^3 * eggs / m^2). This function
implements a linear model to estimate egg density on substrate in
\code{\link{calc_under_index}} \insertCite{HaegeleEtal1979}{SpawnIndex}.
}
\examples{
data(pars)
dens_under_sub(sub_layers = 4, sub_prop = 0.5)
}
\references{
\insertAllCited
}
\seealso{
\code{\link{pars}} \code{\link{calc_under_index}}
\code{\link{dens_under_alg}}

Other calculation functions: 
\code{\link{calc_macro_index}()},
\code{\link{calc_sok_sb}()},
\code{\link{calc_surf_index}()},
\code{\link{calc_under_index}()},
\code{\link{dens_surf}()},
\code{\link{dens_under_alg}()},
\code{\link{eggs_macro}()},
\code{\link{eggs_to_sb}()}
}
\concept{calculation functions}
