% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{egg_dens_surf}
\alias{egg_dens_surf}
\title{Calculate surface spawn egg density}
\usage{
egg_dens_surf(alpha = pars$surface$alpha, beta = pars$surface$beta, egg_lyrs)
}
\description{
Calculate surface spawn egg density
}
