% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{egg_dens_under_alg}
\alias{egg_dens_under_alg}
\title{Calculate for understory spawn egg density on algae}
\usage{
egg_dens_under_alg(
  vartheta = pars$understory$vartheta,
  varrho = pars$understory$varrho,
  varsigma = pars$understory$varsigma,
  alg_layers,
  alg_prop,
  coeff
)
}
\description{
Calculate for understory spawn egg density on algae
}
