% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{eggs_macro}
\alias{eggs_macro}
\title{Calculate Macrocystis spawn number of eggs per plant.}
\usage{
eggs_macro(
  xi = pars$macrocystis$xi,
  gamma = pars$macrocystis$gamma,
  delta = pars$macrocystis$delta,
  epsilon = pars$macrocystis$epsilon,
  egg_layers,
  height,
  stalks_per_plant,
  quiet = FALSE
)
}
\arguments{
\item{xi}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{gamma}{Numeric. Regression exponent on egg layers; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{delta}{Numeric. Regression exponent on plant height; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{epsilon}{Numeric. Regression exponent on number of stalks per plant;
from \code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{egg_layers}{Numeric. Number of egg layers. Message if < 0.}

\item{height}{Numeric. Plant height in metres. Message if < 0.}

\item{stalks_per_plant}{Numeric. Number of stalks per plant. Message if < 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. Number of eggs per plant in thousands. Message if < 0.
}
\description{
Calculate Pacific Herring Macrocystis spawn number of eggs per plant in
thousands of eggs per plant (10^3 * eggs / plant). This function implements a
nonlinear multiple regression model to estimate the number of eggs in
\code{\link{calc_macro_index}}
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.
}
\examples{
data(pars)
eggs_macro(egg_layers = 4, height = 3, stalks_per_plant = 2)
}
\references{
\insertAllCited
}
\seealso{
\code{\link{pars}} \code{\link{calc_macro_index}}

Other calculation functions: 
\code{\link{calc_macro_index}()},
\code{\link{calc_sok_sb}()},
\code{\link{calc_surf_index}()},
\code{\link{calc_under_index}()},
\code{\link{dens_surf}()},
\code{\link{dens_under_alg}()},
\code{\link{dens_under_sub}()},
\code{\link{eggs_to_sb}()}
}
\concept{calculation functions}
