% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{eggs_to_sb}
\alias{eggs_to_sb}
\title{Convert number of eggs to spawning biomass.}
\usage{
eggs_to_sb(
  omega = pars$conversion$omega,
  female = pars$conversion$female,
  quiet = FALSE
)
}
\arguments{
\item{omega}{Numeric. The number of eggs per kilogram of female spawners;
from \code{\link{pars}}. Message if < 0.}

\item{female}{Numeric. The proportion of spawners that are female; from
\code{\link{pars}}. Message if < 0 and/or > 1.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. The conversion factor for eggs to spawn index in tonnes
(i.e., biomass). Divide the number of eggs by the conversion factor to get
biomass. Message if < 0.
}
\description{
Calculate the conversion factor for the number of Pacific Herring eggs to
spawning biomass (i.e., the spawn index) in tonnes.
}
\examples{
data(pars)
eggs_to_sb()
}
\references{
\insertAllCited
}
\seealso{
\code{\link{pars}}

Other calculation functions: 
\code{\link{calc_macro_index}()},
\code{\link{calc_sok_sb}()},
\code{\link{calc_surf_index}()},
\code{\link{calc_under_index}()},
\code{\link{dens_surf}()},
\code{\link{dens_under_alg}()},
\code{\link{dens_under_sub}()},
\code{\link{eggs_macro}()}
}
\concept{calculation functions}
