% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_width}
\alias{get_width}
\title{Load median spawn width.}
\usage{
get_width(where, a = areas)
}
\arguments{
\item{where}{List. Location of the Pacific Herring surface spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{load_area_data}}.}
}
\value{
List with three tables: median region (\code{WidthReg}), section
  (\code{WidthSec}), and pool (\code{WidthPool}) widths in metres (m) for the
  areas in \code{a}.
}
\description{
Load median spawn width in metres (m) for Pacific Herring surface spawn index
calculations. Observed width is not preferred for surface spawn surveys
because surveyors tend to underestimate spawn width
\insertCite{HayKronlund1987}{SpawnIndex}. Instead, the preferred with comes
from underwater surveys \insertCite{GrinnellEtalYYYY}{SpawnIndex}.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = areaLoc)
width_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(
    region_std = "RegionStd", section_std = "SectionStd",
    pool_std = "PoolStd"
  )
)
width_bar <- get_width(where = width_loc, a = areas)
width_bar
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{load_area_data}}
  \code{\link{calc_surf_spawn}}
}
