% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{GetWidth}
\alias{GetWidth}
\title{Load median spawn width.}
\usage{
GetWidth(where, a = areas)
}
\arguments{
\item{where}{List. Location of the Pacific Herring surface spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations. Returned from
\code{\link{LoadAreaData}}.}
}
\value{
Table with median region (WidthReg), section (WidthSec), and pool
  (WidthPool) widths in metres (m) for the areas in \code{a}.
}
\description{
Load median spawn width in metres (m) for Pacific Herring surface spawn index
calculations.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
widthLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(
    regionStd = "RegionStd", sectionStd = "SectionStd", poolStd = "PoolStd"
  )
)
barWidth <- GetWidth(where = widthLoc, a = areas)
barWidth
}
\seealso{
\code{\link{CalcSurfSpawn}} \code{\link{LoadAreaData}}
}
