% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{HerringSections}
\alias{HerringSections}
\title{Pacific Herring Section polygons.}
\format{
Shapefile. Cluster of files that define the spatial boundaries of
Pacific Herring Sections, Statistical Areas, and Regions as polygons.
Coordinate reference system: these polygons are projected in BC Albers (NAD
83).
}
\description{
Pacific Herring Section polygons, with information to aggregate Sections by
Statistical Area and Region.
}
\note{
These are not official Section polygons such as might be used to
ascribe catch to a given area for management purposes; these polygons are
only to be used for making nice maps. Do not use these polygons for
navigation.
}
\examples{
sections_files <- system.file("extdata", "Sections", package = "SpawnIndex")
library(sf)
st_read(dsn = sections_files, layer = "HerringSections", quiet = TRUE)
}
\seealso{
\code{\link{load_sections}}

Other raw data: 
\code{\link{HerringSpawn}}
}
\concept{raw data}
