% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_all_spawn}
\alias{load_all_spawn}
\title{Get the all spawn table.}
\usage{
load_all_spawn(where, a, yrs, ft2m = 0.3048)
}
\arguments{
\item{where}{List. Location of the Pacific Herring understory spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{load_area_data}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{ft2m}{Numeric. Conversion factor for feet to metres.}
}
\value{
Tibble. Contains additional spawn survey data including start and end
  dates, as well as spawn length, width, and depth. Other information in this
  tibble comes from \code{a}: Region, Statistical Area, Section, and Location
  code.
}
\description{
Get the all spawn table.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
all_spawn_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(all_spawn = "tSSAllspawn", stations = "tSSStations")
)
all_spawn <- load_all_spawn(
  where = all_spawn_loc, a = areas, yrs = 2010:2015
)
all_spawn
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{load_area_data}}
}
