% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_area_data}
\alias{load_area_data}
\title{Load Pacific Herring areas.}
\usage{
load_area_data(
  reg,
  sec_sub = NULL,
  where,
  in_crs = "+init=epsg:4326",
  out_crs = "+init=epsg:3005",
  groups = NULL,
  quiet = FALSE
)
}
\arguments{
\item{reg}{Character. Region of interest (i.e., HG, PRD, CC, SoG, WCVI, A27,
or A2W).}

\item{sec_sub}{Numeric vector or NULL Subset of Sections to include in the
analysis, or NULL to include all the Sections in the region.}

\item{where}{List. Location of the Pacific Herring spawn survey database (see
examples).}

\item{in_crs}{Character. Input coordinate reference system;
\href{https://spatialreference.org/}{use EPSG codes if desired}.}

\item{out_crs}{Character. Output coordinate reference system;
\href{https://spatialreference.org/}{use EPSG codes if desired}.}

\item{groups}{Tibble or NULL. Optional table to add a "Group" column to the
results, say to aggregate data by combinations of Sections. Must have a
column named "Group", and one or more of "StatArea", "Section",
"LocationCode". Set to NULL to ignore (and Group column will be NA).}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Tibble. Table of geographic information for Pacific Herring: SAR,
Region, Region name, Statistical Area, Group, Section, Location code,
Location name, Pool, Eastings, Northings, Longitude, and Latitude.
}
\description{
Load Pacific Herring areas. Herring areas are kept in two files: the section
file has coarse area information, and location file has finer details. This
function merges these two files, and drops unnecessary rows and columns. In
addition, 'groups' are created for certain regions based on section numbers.
The output is a data frame with both coarse- and fine-scale area information
for the region(s) in question. There is an option to subset the sections if
desired.
}
\note{
This function requires 32-bit R to load data from the 32-bit MS Access
database.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
areas
secs <- c(231:233, 241, 245)
grps <- tibble::tibble(
  Section = c(231, 232, 233, 241),
  Group = c("Alberni Int", "Barkley", "Barkley", "Tofino Int")
)
areas_sec_grp <- load_area_data(
  reg = "WCVI", where = area_loc, groups = grps, sec_sub = secs
)
dplyr::distinct(dplyr::select(
  areas_sec_grp, Region, StatArea, Group, Section
))
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}}

Other load functions: 
\code{\link{get_width}()},
\code{\link{load_all_spawn}()}
}
\concept{load functions}
