% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{LoadAllSpawn}
\alias{LoadAllSpawn}
\title{Get the all spawn table.}
\usage{
LoadAllSpawn(where, a, yrs, ft2m = 0.3048)
}
\arguments{
\item{where}{List. Location of the Pacific Herring understory spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations; from
\code{\link{LoadAreaData}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{ft2m}{Numeric. Conversion factor for feet to metres.}
}
\value{
Tibble. Contains additional spawn survey data including start and end
  dates, as well as spawn length, width, and depth. Other information in this
  tibble comes from \code{a}: Region, Statistical Area, Section, and Location
  code.
}
\description{
Get the all spawn table.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
allSpawnLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(allSpawn = "tSSAllspawn", stations = "tSSStations")
)
allSpawn <- LoadAllSpawn(where = allSpawnLoc, a = areas, yrs = 2010:2015)
allSpawn
}
\seealso{
\code{\link{LoadAreaData}}
}
