% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_na}
\alias{max_na}
\title{Calculate the maximum.}
\usage{
max_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{omit_na}{Logical. Omit NAs; default is TRUE.}
}
\value{
A numeric vector.
}
\description{
Calculate the maximum; an alternative to \code{max(x)} that returns \code{NA} if \code{x} is
all \code{NA} or an empty set.
}
\examples{
max(NA, na.rm = TRUE)
max_na(NA)
}
\seealso{
\code{\link{max}}

Other utility functions: 
\code{\link{mean_na}()},
\code{\link{paste_nicely}()},
\code{\link{sum_na}()},
\code{\link{unique_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
