% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_na}
\alias{mean_na}
\title{Calculate the mean.}
\usage{
mean_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{omit_na}{Logical. Omit NAs; default is TRUE.}
}
\value{
A numeric vector.
}
\description{
Calculate the mean; an alternative to \code{mean(x)} that returns \code{NA} if \code{x} is
all \code{NA} or an empty set.
}
\examples{
mean(NA, na.rm = TRUE)
mean_na(NA)
}
\seealso{
\code{\link{mean}}

Other utility functions: 
\code{\link{max_na}()},
\code{\link{paste_nicely}()},
\code{\link{sum_na}()},
\code{\link{unique_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
