% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_na}
\alias{min_na}
\title{Minimum
Calculate minimum if there are non-NA values, return NA if all values are NA}
\usage{
min_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{Object.}

\item{omit_na}{Omit NAs.}
}
\value{
Numeric.
}
\description{
Minimum
Calculate minimum if there are non-NA values, return NA if all values are NA
}
\seealso{
\code{\link{max_na}}, \code{\link{sum_na}}, \code{\link{mean_na}},
\code{\link{wt_mean_na}}, \code{\link{roll_mean_na}},
\code{\link{unique_na}}

Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{paste_nicely}()},
\code{\link{roll_mean_na}()},
\code{\link{sum_na}()},
\code{\link{unique_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
