% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{num_eggs_macro}
\alias{num_eggs_macro}
\title{Calculate Macrocystis spawn number of eggs per plant}
\usage{
num_eggs_macro(
  xi = pars$macrocystis$xi,
  gamma = pars$macrocystis$gamma,
  delta = pars$macrocystis$delta,
  epsilon = pars$macrocystis$epsilon,
  egg_layers,
  height,
  stalks_per_plant
)
}
\description{
Calculate Macrocystis spawn number of eggs per plant
}
