% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_nicely}
\alias{paste_nicely}
\title{Paste things nicely.}
\usage{
paste_nicely(
  x,
  int_char = ", ",
  n_char = "and ",
  do_quotes = FALSE,
  quotes = c("`", "`")
)
}
\arguments{
\item{x}{Objects to paste.}

\item{int_char}{Character to put between objects. Default is \code{", "}.}

\item{n_char}{Character to put between last two objects. Default is \code{"and "}.}

\item{do_quotes}{Logical Add quotations. Default is FALSE (no quotes).}

\item{quotes}{Character vector of length two (i.e., preceding and following
quotes). Default is c("\code{", "}").}
}
\value{
Character. Objects pasted together.
}
\description{
Paste things together nicely for printing. Quotes and oxford comma included.
}
\examples{
paste_nicely(x = letters[1:3])
}
\seealso{
Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{sum_na}()},
\code{\link{unique_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
