% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roll_mean_na}
\alias{roll_mean_na}
\title{Rolling mean (window)
Fill in NA values with rolling mean of previous values}
\usage{
roll_mean_na(dat, n, omit_na = TRUE)
}
\arguments{
\item{dat}{Object.}

\item{n}{Window size.}

\item{omit_na}{Omit NAs.}
}
\value{
Numeric.
}
\description{
Rolling mean (window)
Fill in NA values with rolling mean of previous values
}
\seealso{
\code{\link{min_na}}, \code{\link{max_na}}, \code{\link{sum_na}},
\code{\link{mean_na}}, \code{\link{wt_mean_na}}, \code{\link{unique_na}}

Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{min_na}()},
\code{\link{paste_nicely}()},
\code{\link{sum_na}()},
\code{\link{unique_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
