% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sections}
\alias{sections}
\title{Pacific Herring Section polygons.}
\format{
Simple feature collection of polygons that define the spatial
boundaries of Pacific Herring Sections, Statistical Areas, and Regions.
Coordinate reference system: these polygons are projected in BC Albers (NAD
83).
}
\usage{
sections
}
\description{
Pacific Herring Section polygons, with information to aggregate Sections by
Statistical Area and Region.
}
\note{
These are not official Section polygons such as might be used to
ascribe catch to a given area for management purposes; these polygons are
only to be used for making nice maps. Do not use these polygons for
navigation.
}
\examples{
data(sections)
sections
}
\seealso{
\code{\link{load_sections}}

Other data: 
\code{\link{algae_coefs}},
\code{\link{intensity}},
\code{\link{pars}},
\code{\link{regions}},
\code{\link{under_width_facs}}
}
\concept{data}
\keyword{datasets}
