% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{under_width_fac}
\alias{under_width_fac}
\title{Understory spawn width correction factors.}
\format{
Tibble with rows for years and columns for regions.
}
\usage{
under_width_fac
}
\description{
Spawn width correction factors for Pacific Herring understory spawn surveys
by region and year. These correction factors account for an error that
resulted in underestimated spawn width for understory dive surveys
\insertCite{ClearyEtal2017,GrinnellEtalYYYY}{SpawnIndex}.
}
\note{
Adjust spawn width for affected understory spawn surveys by multiplying
  the observed width by the corresponding correction factor for the affected
  years.
}
\examples{
data(under_width_fac)
under_width_fac
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CalcUnderSpawn}}
}
\keyword{datasets}
