% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{underWidthFac}
\alias{underWidthFac}
\title{Understory spawn width correction factors.}
\format{Tibble with 12 rows and 8 columns (one for year, and one for each
regions).}
\usage{
data(underWidthFac)
}
\description{
Spawn width correction factors for Pacific Herring understory spawn surveys
by region and year.
}
\note{
Adjust spawn width for affected understory spawn surveys by multiplying
  the observed width by the corresponding correction fator for the affected
  years.
}
\examples{
data(underWidthFac)
}
\seealso{
\code{\link{CalcUnderSpawn}}
}
\keyword{datasets}
