% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_na}
\alias{unique_na}
\title{Unique
Calculate unique if there are non-NA values, return NA if all values are NA}
\usage{
unique_na(x)
}
\arguments{
\item{x}{Vector to consider.}

\item{omit_na}{Omit NAs? Logical, default TRUE.}
}
\value{
Numeric.
}
\description{
Unique
Calculate unique if there are non-NA values, return NA if all values are NA
}
\seealso{
Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{min_na}()},
\code{\link{paste_nicely}()},
\code{\link{roll_mean_na}()},
\code{\link{sum_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
