% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wt_mean_na}
\alias{wt_mean_na}
\title{Calculate the weighted mean.}
\usage{
wt_mean_na(x, w, omit_na = TRUE)
}
\arguments{
\item{x}{A numeric vector of values whose weighted mean is to be computed.}

\item{w}{A numeric vector of weights of the same length as \code{x}.}

\item{omit_na}{Logical. Omit NAs; default is TRUE.}
}
\value{
A numeric vector.
}
\description{
Calculate the weighted mean; an alternative to \code{weighted.mean(x)} that
returns \code{NA} if \code{x} is all \code{NA} or an empty set.
}
\examples{
weighted.mean(NA, na.rm = TRUE)
wt_mean_na(NA, w = 1)
}
\seealso{
\code{\link{weighted.mean}}

Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{paste_nicely}()},
\code{\link{sum_na}()},
\code{\link{unique_na}()}
}
\concept{utility functions}
