% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{algae_coefs}
\alias{algae_coefs}
\title{Algae types and coefficients.}
\format{
Tibble with 8 rows and 3 columns: algae type, algae code, and
coefficient.
}
\usage{
algae_coefs
}
\description{
Algae types and coefficients for Pacific Herring understory spawn surveys
\insertCite{Schweigert2005}{SpawnIndex}. Algae coefficients account for the
effect of algae morphology on Pacific Herring egg density
\insertCite{GrinnellEtalYYYY}{SpawnIndex}.
}
\examples{
data(algae_coefs)
algae_coefs
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{calc_under_index}}

Other data: 
\code{\link{intensity}},
\code{\link{pars}},
\code{\link{regions}},
\code{\link{sections}},
\code{\link{under_width_facs}}
}
\concept{data}
\keyword{datasets}
