% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_biomass_sok}
\alias{calc_biomass_sok}
\title{Calculate spawning biomass from spawn-on-kelp (SOK) harvest.}
\usage{
calc_biomass_sok(
  sok,
  nu = pars$sok$nu,
  upsilon = pars$sok$upsilon,
  w = pars$sok$w,
  theta = calc_egg_conversion(),
  quiet = FALSE
)
}
\arguments{
\item{sok}{Numeric. Weight of spawn-on-kelp (SOK) harvest in kilograms.
Message if < 0.}

\item{nu}{Numeric. Proportion of SOK product that is kelp; from
\code{\link{pars}}. Message if < 0 and/or > 1.}

\item{upsilon}{Numeric. SOK product weight increase due to brining as a
proportion; from \code{\link{pars}}. Message if < 0 and/or > 1.}

\item{w}{Numeric. Average weight in kilograms of a fertilized egg; from
\code{\link{pars}}. Message if < 0.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{calc_egg_conversion}}. Message if < 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. Spawning biomass in tonnes. Message if < 0.
}
\description{
Calculate spawning biomass in tonnes from spawn-on-kelp (SOK) harvest in
kilograms.
}
\examples{
data(pars)
calc_biomass_sok(sok = 100)
}
\references{
\insertAllCited
}
\seealso{
\code{\link{calc_egg_conversion}} \code{\link{pars}}

Other calculation functions: 
\code{\link{calc_egg_conversion}()},
\code{\link{calc_macro_spawn}()},
\code{\link{calc_surf_spawn}()},
\code{\link{calc_under_spawn}()}
}
\concept{calculation functions}
