% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calc_egg_conversion}
\alias{calc_egg_conversion}
\title{Convert number of eggs to spawning biomass.}
\usage{
calc_egg_conversion(
  omega = pars$conversion$omega,
  female = pars$conversion$female,
  quiet = FALSE
)
}
\arguments{
\item{omega}{Numeric. The number of eggs per kilogram of female spawners;
from \code{\link{pars}}. Message if < 0.}

\item{female}{Numeric. The proportion of spawners that are female; from
\code{\link{pars}}. Message if not in [0, 1].}

\item{quiet}{Logical. Print messages; default is FALSE.}
}
\value{
Numeric. The conversion factor for eggs to spawn index in tonnes
  (i.e., biomass). Divide the number of eggs by the conversion factor to get
  biomass. Message if < 0.
}
\description{
Calculate the conversion factor for the number of Pacific Herring eggs to the
spawn index (i.e., biomass) in tonnes.
}
\examples{
data(pars)
calc_egg_conversion()
}
\references{
\insertAllCited
}
\seealso{
\code{\link{pars}}

Other calculation functions: 
\code{\link{calc_biomass_sok}()},
\code{\link{calc_macro_spawn}()},
\code{\link{calc_surf_spawn}()},
\code{\link{calc_under_spawn}()}
}
\concept{calculation functions}
