% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcBiomassSOK}
\alias{CalcBiomassSOK}
\title{Calculate spawning biomass from spawn-on-kelp (SOK) harvest.}
\usage{
CalcBiomassSOK(
  SOK,
  eggKelpProp = 0.88,
  eggBrineProp = 1/1.13,
  eggWt = 2.38 * 10^-6,
  f
)
}
\arguments{
\item{SOK}{Weight of spawn-on-kelp (SOK) harvest in kilograms.}

\item{eggKelpProp}{Proportion of SOK product that is eggs, not kelp (default
0.88) \insertCite{ShieldsEtal1985}{SpawnIndex}.}

\item{eggBrineProp}{Pproportion of SOK product that is eggs after brining
(default 1/1.13) \insertCite{WhyteEnglar1977}{SpawnIndex}.}

\item{eggWt}{Average weight in kilograms of a fertilized egg (default
2.38*10^-6) \insertCite{HayMiller1982}{SpawnIndex}.}

\item{f}{Egg conversion factor (eggs to biomass) from
\code{\link{CalcEggConversion}}.}
}
\value{
Numeric. Spawning biomass in tonnes.
}
\description{
Calculate spawning biomass in tonnes from spawn-on-kelp (SOK) harvest in
kilograms.
}
\examples{
ECF <- CalcEggConversion( )
CalcBiomassSOK( SOK=100, f=ECF )
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CalcEggConversion}}
}
