% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcBiomassSOK}
\alias{CalcBiomassSOK}
\title{Calculate spawning biomass from spawn-on-kelp (SOK) harvest.}
\usage{
CalcBiomassSOK(
  SOK,
  nu = pars$SOK$nu,
  upsilon = pars$SOK$upsilon,
  M = pars$SOK$M,
  theta
)
}
\arguments{
\item{SOK}{Numeric. Weight of spawn-on-kelp (SOK) harvest in kilograms.}

\item{nu}{Numeric. Proportion of SOK product that is eggs, not kelp
\insertCite{ShieldsEtal1985}{SpawnIndex}.}

\item{upsilon}{Numeric. Pproportion of SOK product that is eggs after
brining \insertCite{WhyteEnglar1977}{SpawnIndex}.}

\item{M}{Numeric. Average weight in kilograms of a fertilized egg
\insertCite{HayMiller1982}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass) from
\code{\link{CalcEggConversion}}.}
}
\value{
Numeric. Spawning biomass in tonnes.
}
\description{
Calculate spawning biomass in tonnes from spawn-on-kelp (SOK) harvest in
kilograms.
}
\examples{
CalcBiomassSOK(SOK = 100, theta = CalcEggConversion())
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CalcEggConversion}} \code{\link{pars}}
}
