% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcEggConversion}
\alias{CalcEggConversion}
\title{Convert number of eggs to spawn index in tonnes.}
\usage{
CalcEggConversion(fecundity = 2e+05, pFemale = 0.5)
}
\arguments{
\item{fecundity}{The number of eggs per kilogram of female spawners (default
200000).}

\item{pFemale}{The proportion of spawners that are female (default 0.5).}
}
\value{
Numeric. The conversion factor for eggs to spawn index in tonnes
(i.e., biomass). Divide the number of eggs by the conversion factor to get
biomass.
}
\description{
Calculate the conversion factor for Pacific Herring, to convert the number of
eggs to spawn index (i.e., biomass) in tonnes.
}
\references{
\insertRef{Hay1985}{SpawnIndex}

\insertRef{HayBrett1988}{SpawnIndex}
}
