% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcEggConversion}
\alias{CalcEggConversion}
\title{Convert number of eggs to spawning biomass.}
\usage{
CalcEggConversion(omega = pars$conversion$omega, phi = pars$conversion$phi)
}
\arguments{
\item{omega}{Numeric. The number of eggs per kilogram of female spawners;
from \code{\link{pars}}.}

\item{phi}{Numeric. The proportion of spawners that are female; from
\code{\link{pars}}.}
}
\value{
Numeric. The conversion factor for eggs to spawn index in tonnes
  (i.e., biomass). Divide the number of eggs by the conversion factor to get
  biomass.
}
\description{
Calculate the conversion factor for the number of Pacific Herring eggs to the
spawn index (i.e., biomass) in tonnes.
}
\examples{
data(pars)
CalcEggConversion()
}
\seealso{
\code{\link{pars}}
}
