% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcSurfSpawn}
\alias{CalcSurfSpawn}
\title{Calculate the surface spawn index.}
\usage{
CalcSurfSpawn(
  where,
  a,
  yrs,
  intYrs = yrs[yrs < 1979],
  rsYrs = intYrs[intYrs < 1951],
  f
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring surface spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations. Returned from
\code{\link{LoadAreaData}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{intYrs}{Numeric vector. Years where intensity categores are used to
determine egg layers (default yrs[yrs < 1979]).}

\item{rsYrs}{Numeric vector. Years where intensity needs to be re-scaled from
5 to 9 categories (default intYrs[intYrs < 1951]).}

\item{f}{Numeric. Egg conversion factor (eggs to biomass) from
\code{\link{CalcEggConversion}}.}
}
\value{
List. The element `SI` is a tibble with surface spawn index in tonnes
  by spawn number and year. The spawn number is the finest spatial scale at
  which we calculate the spawn index. Other information in this tibble comes
  from `a`: Region, Statistical Area, Section, and Location code.
}
\description{
Calculate the Pacific Herring surface spawn index in tonnes. Note that the
'spawn index' is a relative index of spawning biomass.
}
\examples{
areaLoc <- list(
  loc = "data", db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
surfLoc <- list(
  loc = "data", db = "HerringSpawn.mdb",
  fns = list(
    regionStd = "RegionStd", sectionStd = "SectionStd", poolStd = "PoolStd",
    surface = "tSSSurface", intensity = "Intensity",
    allSpawn = "tSSAllspawn"
  )
)
surfSpawn <- CalcSurfSpawn(
  where = surfLoc, a = areas, yrs = 2010:2015,
  f = CalcEggConversion()
)
surfSpawn$SI
}
\references{
\insertRef{SchweigertEtal1997}{SpawnIndex}
}
\seealso{
\code{\link{LoadAreaData}} \code{\link{CalcEggConversion}}
}
