% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcSurfSpawn}
\alias{CalcSurfSpawn}
\title{Calculate the surface spawn index.}
\usage{
CalcSurfSpawn(
  where,
  a,
  widths,
  yrs,
  intense = intensity,
  intYrs = yrs[yrs < 1979],
  rsYrs = intYrs[intYrs < 1951],
  alpha = pars$surface$alpha,
  beta = pars$surface$beta,
  theta = CalcEggConversion()
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring surface spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations. Returned from
\code{\link{LoadAreaData}}.}

\item{widths}{Tibble. Table of median region, section, and pool widths in
metres (m). Returned from \code{\link{GetWidth}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{intense}{Tibble. Table of spawn intensity categories and number of egg
layers; from \code{\link{intensity}}.}

\item{intYrs}{Numeric vector. Years where intensity categores are used to
determine egg layers.}

\item{rsYrs}{Numeric vector. Years where intensity needs to be re-scaled from
5 to 9 categories.}

\item{alpha}{Numeric. Regression intercept; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{beta}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{CalcEggConversion}}.}
}
\value{
List. The element `SI` is a tibble with surface spawn index
  (`SurfSI`) in tonnes by spawn number and year. The spawn number is the
  finest spatial scale at which we calculate the spawn index. Other
  information in this tibble comes from `a`: Region, Statistical Area,
  Section, and Location code.
}
\description{
Calculate the Pacific Herring surface spawn index in tonnes.
}
\note{
The 'spawn index' is a relative index of spawning biomass.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
widthLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(
    regionStd = "RegionStd", sectionStd = "SectionStd", poolStd = "PoolStd"
  )
)
barWidth <- GetWidth(where = widthLoc, a = areas)
data(pars)
data(intensity)
surfLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(surface = "tSSSurface", allSpawn = "tSSAllspawn")
)
surfSpawn <- CalcSurfSpawn(
  where = surfLoc, a = areas, widths = barWidth, yrs = 2010:2015
)
surfSpawn$SI
}
\references{
\insertAllCited
}
\seealso{
\code{\link{LoadAreaData}} \code{\link{GetWidth}}
  \code{\link{CalcEggConversion}} \code{\link{pars}} \code{\link{intensity}}
}
