% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcUnderSpawn}
\alias{CalcUnderSpawn}
\title{Calculate the understory spawn index.}
\usage{
CalcUnderSpawn(
  where,
  a,
  yrs,
  algaeCoefs = algaeCoefs,
  tau = underWidthFac,
  alpha = pars$understory$alpha,
  beta = pars$understory$beta,
  gamma = pars$under$gamma,
  delta = pars$understory$delta,
  theta = CalcEggConversion()
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring understory spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations. Returned from
\code{\link{LoadAreaData}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{algaeCoefs}{Tibble. Table of algae coefficients; from
\code{\link{algaeCoefs}}.}

\item{tau}{Tibble. Table of understory spawn width adjustment factors from
\code{\link{underWidthFac}}.}

\item{alpha}{Numeric. Regression slope for substrate; from \code{\link{pars}}
\insertCite{HaegeleEtal1979}{SpawnIndex}.}

\item{beta}{Numeric. Regression slope for algae; from \code{\link{pars}}
\insertCite{Schweigert2005}{SpawnIndex}.}

\item{gamma}{Numeric. Regression exponent on number of egg layers; from
\code{\link{pars}} \insertCite{Schweigert2005}{SpawnIndex}.}

\item{delta}{Numeric. Regression exponent on proportion of algae; from
\code{\link{pars}} \insertCite{Schweigert2005}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{CalcEggConversion}}.}
}
\value{
List. The element `SI` is a tibble with understory spawn index
  (`UnderSI`) in tonnes by spawn number and year. The spawn number is the
  finest spatial scale at which we calculate the spawn index. Other
  information in this tibble comes from `a`: Region, Statistical Area,
  Section, and Location code.
}
\description{
Calculate the Pacific Herring understory spawn index in tonnes.
}
\note{
The 'spawn index' is a relative index of spawning biomass.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
underLoc <- list(
  loc=file.path(dbLoc), db="HerringSpawn.mdb",
  fns=list(allSpawn="tSSAllspawn", algTrans="tSSVegTrans",
           stations="tSSStations", algae="tSSVegetation",
           typeAlg="tSSTypeVegetation") )
data(underWidthFac)
data(pars)
data(algaeCoefs)
underSpawn <- CalcUnderSpawn(
  where = underLoc, a = areas, yrs = 2010:2015
)
underSpawn$SI
}
\references{
\insertAllCited
}
\seealso{
\code{\link{LoadAreaData}} \code{\link{CalcEggConversion}}
  \code{\link{pars}} \code{\link{algaeCoefs}}
}
