% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcUnderSpawn}
\alias{CalcUnderSpawn}
\title{Calculate the understory spawn index.}
\usage{
CalcUnderSpawn(
  where,
  a,
  yrs,
  algCoefs = algae_coefs,
  tau = under_width_fac,
  varphi = pars$understory$varphi,
  vartheta = pars$understory$vartheta,
  varrho = pars$under$varrho,
  varsigma = pars$understory$varsigma,
  theta = CalcEggConversion()
)
}
\arguments{
\item{where}{List. Location of the Pacific Herring understory spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{LoadAreaData}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{algCoefs}{Tibble. Table of algae coefficients; from
\code{\link{algae_coefs}}.}

\item{tau}{Tibble. Table of understory spawn width adjustment factors from
\code{\link{under_width_fac}}.}

\item{varphi}{Numeric. Regression slope for substrate; from
\code{\link{pars}} \insertCite{HaegeleEtal1979}{SpawnIndex}.}

\item{vartheta}{Numeric. Regression slope for algae; from \code{\link{pars}}
\insertCite{Schweigert2005}{SpawnIndex}.}

\item{varrho}{Numeric. Regression exponent on number of egg layers; from
\code{\link{pars}} \insertCite{Schweigert2005}{SpawnIndex}.}

\item{varsigma}{Numeric. Regression exponent on proportion of algae; from
\code{\link{pars}} \insertCite{Schweigert2005}{SpawnIndex}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass); from
\code{\link{CalcEggConversion}}.}
}
\value{
List. The element \code{SI} is a tibble with understory spawn index
  (\code{UnderSI}) in tonnes by spawn number and year. The spawn number is
  the finest spatial scale at which we calculate the spawn index. Other
  information in this tibble comes from \code{a}: Region, Statistical Area,
  Section, and Location code.
}
\description{
Calculate the Pacific Herring understory spawn index in tonnes.
}
\note{
The `spawn index' is a relative index of spawning biomass.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
underLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(
    allSpawn = "tSSAllspawn", algTrans = "tSSVegTrans",
    stations = "tSSStations", algae = "tSSVegetation"
  )
)
data(under_width_fac)
data(pars)
data(algae_coefs)
underSpawn <- CalcUnderSpawn(where = underLoc, a = areas, yrs = 2010:2015)
underSpawn$SI
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{LoadAreaData}}
  \code{\link{CalcEggConversion}} \code{\link{pars}} \code{\link{algae_coefs}}
}
