% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{CalcUnderSpawn}
\alias{CalcUnderSpawn}
\title{Calculate the understory spawn index.}
\usage{
CalcUnderSpawn(where, a, yrs, tau = underWidthFac, theta)
}
\arguments{
\item{where}{List. Location of the Pacific Herring understory spawn database
(see examples).}

\item{a}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to inlude in calculations. Returned from
\code{\link{LoadAreaData}}.}

\item{yrs}{Numeric vector. Years(s) to include in the calculations, usually
staring in 1951.}

\item{tau}{Tibble. Table of understory spawn width adjustment factors from
\code{\link{underWidthFac}}.}

\item{theta}{Numeric. Egg conversion factor (eggs to biomass) from
\code{\link{CalcEggConversion}}.}
}
\value{
List. The element `SI` is a tibble with understory spawn index
  (`UnderSI`) in tonnes by spawn number and year. The spawn number is the
  finest spatial scale at which we calculate the spawn index. Other
  information in this tibble comes from `a`: Region, Statistical Area,
  Section, and Location code.
}
\description{
Calculate the Pacific Herring understory spawn index in tonnes.
}
\note{
The 'spawn index' is a relative index of spawning biomass.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- LoadAreaData(reg = "WCVI", where = areaLoc)
underLoc <- list(
  loc=file.path(dbLoc), db="HerringSpawn.mdb",
  fns=list(allSpawn="tSSAllspawn", algTrans="tSSVegTrans",
           stations="tSSStations", algae="tSSVegetation",
           typeAlg="tSSTypeVegetation") )
data(underWidthFac)
underSpawn <- CalcUnderSpawn(
  where = underLoc, a = areas, yrs = 2010:2015,
  theta = CalcEggConversion()
)
underSpawn$SI
}
\references{
\insertRef{}{SpawnIndex}
}
\seealso{
\code{\link{LoadAreaData}} \code{\link{CalcEggConversion}}
  \code{\link{pars}}
}
