% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_numeric}
\alias{check_numeric}
\title{Check values for NA or not numeric.}
\usage{
check_numeric(dat, quiet = FALSE)
}
\arguments{
\item{dat}{List. Message if NAs present; error if is not numeric.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
NULL.
}
\description{
Check values for NA or not numeric. Gives a message if there are NA(s), and
and error if it's not numeric.
}
\examples{
alpha <- c(1:3, NA)
beta <- "a"
check_numeric(dat = list(alpha = alpha, beta = beta))
}
\seealso{
Other helper functions: 
\code{\link{check_tibble}()}
}
\concept{helper functions}
