% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_numeric}
\alias{check_numeric}
\title{Check that an object is numeric.}
\usage{
check_numeric(dat, quiet = FALSE)
}
\arguments{
\item{dat}{List of objects. Message if NAs; error if not numeric.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
NULL.
}
\description{
Check that an object is numeric and if there are NAs. Message if NAs, and
error if not numeric.
}
\examples{
x <- c(1:3, NA)
y <- "a"
check_numeric(dat = list(x = x, y = y))
}
\seealso{
Other checker functions: 
\code{\link{check_tibble}()}
}
\concept{checker functions}
