% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_tibble}
\alias{check_tibble}
\title{Check that an object is a tibble.}
\usage{
check_tibble(dat, min_rows = 1, quiet = FALSE)
}
\arguments{
\item{dat}{List of objects. Message if fewer rows; error if not a tibble.}

\item{min_rows}{Numeric. Minimum number of rows in the tibble; default is 1.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
NULL.
}
\description{
Check that an object is a tibble and has a minimum number of rows. Message if
fewer rows, and error if not a tibble.
}
\examples{
x <- data.frame(a = 1, b = 2)
y <- tibble()
check_tibble(dat = list(x = x, y = y))
}
\seealso{
Other checker functions: 
\code{\link{check_numeric}()}
}
\concept{checker functions}
