% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{egg_dens_surf}
\alias{egg_dens_surf}
\title{Calculate surface spawn egg density.}
\usage{
egg_dens_surf(
  alpha = pars$surface$alpha,
  beta = pars$surface$beta,
  egg_layers,
  quiet = FALSE
)
}
\arguments{
\item{alpha}{Numeric. Regression intercept; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{beta}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.}

\item{egg_layers}{Numeric. Number of egg layers. Message if <= 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. Egg density in thousands of eggs per square metre. Message
if <= 0.
}
\description{
Calculate Pacific Herring surface spawn egg density in thousands of eggs per
square metre (10^3 * eggs / m^2). This function implements a linear
regression model to estimate egg density in \code{\link{calc_surf_spawn}}
\insertCite{SchweigertEtal1997}{SpawnIndex}.
}
\note{
There is an error in the original report
\insertCite{SchweigertEtal1997}{SpawnIndex}; surface egg density is in
thousands of eggs per square metre.
}
\examples{
data(pars)
egg_dens_surf(egg_layers = 4)
}
\references{
\insertAllCited
}
\seealso{
\code{\link{pars}} \code{\link{calc_surf_spawn}}

Other calculation functions: 
\code{\link{calc_biomass_sok}()},
\code{\link{calc_egg_conversion}()},
\code{\link{calc_macro_spawn}()},
\code{\link{calc_surf_spawn}()},
\code{\link{calc_under_spawn}()},
\code{\link{egg_dens_under_alg}()},
\code{\link{egg_dens_under_sub}()},
\code{\link{num_eggs_macro}()}
}
\concept{calculation functions}
