% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{egg_dens_under_sub}
\alias{egg_dens_under_sub}
\title{Calculate understory spawn egg density on substrate}
\usage{
egg_dens_under_sub(varphi = pars$understory$varphi, sub_lyrs, sub_prop)
}
\description{
Calculate understory spawn egg density on substrate
}
