% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{HerringSpawn}
\alias{HerringSpawn}
\title{Example Pacific Herring spawn survey database.}
\format{MS Access databases with tables:
\describe{
  \item{Location}{Spatial information for Locations.}
  \item{Sections}{Spatial information for Sections.}
  \item{PoolStd}{Spawn width for Beds.}
  \item{SectionStd}{Spawn width for Sections.}
  \item{RegionStd}{Spawn width for Regions.}
  \item{tSSAllspawn}{Spawn survey information.}
  \item{tSSMacPlant}{Macrocystis plant information.}
  \item{tSSMacTrans}{Macrocystis transect information.}
  \item{tSSStations}{Understory quadrat information.}
  \item{tSSSurface}{Surface spawn information}
  \item{tSSVegetation}{Understory spawn algae information.}
  \item{tSSVegTrans}{Understory spawn transect information.}
}}
\description{
Example Pacific Herring spawn survey database. This database contains a
subset of tables from the spawn survey database, and some additional tables
from other databases to ensure the examples work.
}
\note{
This is a 32-bit MS Access database, and it requires 32-bit R to access
  the data. In addition, MS Windows is required to access the data using the
  RODBC package. This example database only includes WCVI from 2010 to 2015.
}
\examples{
dbLoc <- system.file("extdata", "HerringSpawn.mdb", package = "SpawnIndex")
dbLoc
\dontrun{
accessDB <- RODBC::odbcConnectAccess(access.file = dbLoc)
hTables <- RODBC::sqlTables(channel = accessDB)
tibble::as_tibble(hTables)
RODBC::odbcClose(channel = accessDB)
}
}
\seealso{
\code{\link{LoadAreaData}} \code{\link{LoadAllSpawn}}
  \code{\link{GetWidth}} \code{\link{CalcSurfSpawn}}
  \code{\link{CalcMacroSpawn}} \code{\link{CalcUnderSpawn}}
}
