% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intensity}
\alias{intensity}
\title{Spawn intensity categories and number of egg layers.}
\format{Tibble with 9 rows and 3 columns: intensity category, description,
  and number of egg layers.}
\usage{
data(intensity)
}
\description{
Spawn intensity categories and number of egg layers for Pacific Herring
surface spawn surveys
\insertCite{HayKronlund1987,SchweigertStoker1988}{SpawnIndex}.
}
\examples{
data(intensity)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CalcSurfSpawn}}
}
\keyword{datasets}
