% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_all_spawn}
\alias{load_all_spawn}
\title{Load the all spawn table.}
\usage{
load_all_spawn(where, areas, years, ft2m = 0.3048, quiet = FALSE)
}
\arguments{
\item{where}{List. Location and name of the Pacific Herring spawn survey
database, as well as required tables (see examples).}

\item{areas}{Tibble. Table of geographic information indicating the subset of
spawn survey observations to include in calculations; from
\code{\link{load_area_data}}.}

\item{years}{Numeric vector. Years(s) to include in the calculations. Message
if < \code{pars$years$assess}.}

\item{ft2m}{Numeric. Conversion factor for feet to metres; default is 0.3048.
Message if not 0.3048.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Tibble. Contains additional spawn survey data including start and end
dates, as well as spawn length, width, and depth. Other information in this
tibble comes from \code{a}: Region, Statistical Area, Section, and Location
code.
}
\description{
Load the all spawn table.
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
all_spawn_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(all_spawn = "tSSAllspawn", stations = "tSSStations")
)
all_spawn <- load_all_spawn(
  where = all_spawn_loc, areas = areas, years = 2010:2015
)
all_spawn
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSpawn}} \code{\link{load_area_data}}

Other load functions: 
\code{\link{load_area_data}()},
\code{\link{load_width}()}
}
\concept{load functions}
