% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_sections}
\alias{load_sections}
\title{Load and wrangle Pacific Herring Section shapefiles.}
\usage{
load_sections(where, areas, buffer = 5000, quiet = FALSE)
}
\arguments{
\item{where}{List. Location and name of the Pacific Herring spawn survey
database, as well as required tables (see examples).}

\item{areas}{Tibble. Table of geographic information indicating the subset of
observations to include; from \code{\link{load_area_data}}.}

\item{buffer}{Numeric. Buffer around polygons; distance in metres. Default
5000.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
List of spatial objects showing Section, Group, Statistical Area, and Region
boundaries.
}
\description{
Load and wrangle Pacific Herring Section shapefiles, and aggregate to
Statistical Areas and Region(s).
}
\examples{
db_loc <- system.file("extdata", package = "SpawnIndex")
area_loc <- list(
  loc = db_loc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
areas <- load_area_data(reg = "WCVI", where = area_loc)
sections_files <- system.file("extdata", "Sections", package = "SpawnIndex")
sections_loc <- list(
  loc_sec = sections_files,
  fns = list(sections = "HerringSections")
)
load_sections(where = sections_loc, areas = areas)
}
\references{
\insertAllCited
}
\seealso{
\code{\link{HerringSections}}

Other load functions: 
\code{\link{load_all_spawn}()},
\code{\link{load_area_data}()},
\code{\link{load_width}()}
}
\concept{load functions}
