% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{LoadAreaData}
\alias{LoadAreaData}
\title{Load Pacific Herring areas.}
\usage{
LoadAreaData(
  reg,
  secSub = NA,
  where,
  inCRS = "+init=epsg:4326",
  outCRS = "+init=epsg:3005"
)
}
\arguments{
\item{reg}{Character. Region of interest (i.e., HG, PRD, CC, SoG, WCVI, A27,
or A2W).}

\item{secSub}{Numeric vector or NA. Subset of Sections to include in the
analysis, or NA to include all the Sections in the region.}

\item{where}{List. Location of the Pacfic Herring "locations" database (see
examples).}

\item{inCRS}{Chracter. Input coordinate reference system;
\href{https://spatialreference.org/}{use EPSG codes if desired}.}

\item{outCRS}{Chracter. Output coordinate reference system;
\href{https://spatialreference.org/}{use EPSG codes if desired}.}
}
\value{
Tibble. Table of geographic information for Pacific Herring: SAR,
  Region, Region name, Statistical Area, Group, Section, Location code,
  Location name, Bed, Eastings, Northings, Longitude, and Latitute.
}
\description{
Load Pacific Herring areas. Herring areas are kept in two files: the section
file has coarse area information, and location file has finer details. This
function merges these two files, and drops unnecessary rows and columns. In
addition, 'groups' are created for certain regions based on section numbers.
The output is a data frame with both coarse- and fine-scale area information
for the region(s) in question. There is an option to subset the sections if
desired.
}
\note{
This function requires 32-bit R to load data from the 32-bit MS Access
  database.
}
\examples{
dbLoc <- system.file("extdata", package = "SpawnIndex")
areaLoc <- list(
  loc = dbLoc, db = "HerringSpawn.mdb",
  fns = list(sections = "Sections", locations = "Location")
)
LoadAreaData(reg = "WCVI", where = areaLoc)
}
