% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{num_eggs_macro}
\alias{num_eggs_macro}
\title{Calculate Macrocystis spawn number of eggs per plant.}
\usage{
num_eggs_macro(
  xi = pars$macrocystis$xi,
  gamma = pars$macrocystis$gamma,
  delta = pars$macrocystis$delta,
  epsilon = pars$macrocystis$epsilon,
  egg_layers,
  height,
  stalks_per_plant,
  quiet = FALSE
)
}
\arguments{
\item{xi}{Numeric. Regression slope; from \code{\link{pars}}
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{gamma}{Numeric. Regression exponent on egg layers; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{delta}{Numeric. Regression exponent on plant height; from
\code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{epsilon}{Numeric. Regression exponent on number of stalks per plant;
from \code{\link{pars}} \insertCite{HaegeleSchweigert1990}{SpawnIndex}.}

\item{egg_layers}{Numeric. Number of egg layers. Message if <= 0.}

\item{height}{Numeric. Plant height in metres. Message if <= 0.}

\item{stalks_per_plant}{Numeric. Number of stalks per plant. Message if <= 0.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Numeric. Number of eggs per plant in thousands. Message if <= 0.
}
\description{
Calculate Pacific Herring Macrocystis spawn number of eggs per plant in
thousands of eggs per plant (10^3 * eggs / plant)
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.
}
\examples{
data(pars)
num_eggs_macro(egg_layers = 4, height = 3, stalks_per_plant = 2)
}
\references{
\insertAllCited
}
\seealso{
\code{\link{pars}} \code{\link{calc_macro_spawn}}

Other calculation functions: 
\code{\link{calc_biomass_sok}()},
\code{\link{calc_egg_conversion}()},
\code{\link{calc_macro_spawn}()},
\code{\link{calc_surf_spawn}()},
\code{\link{calc_under_spawn}()},
\code{\link{egg_dens_surf}()},
\code{\link{egg_dens_under_alg}()},
\code{\link{egg_dens_under_sub}()}
}
\concept{calculation functions}
