% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pars}
\alias{pars}
\title{Parameters for spawn index calculations}
\format{List with 5 items:
\describe{
  \item{conversion}{List with 2 items:
    omega is the number of eggs per kilogram of female spawners
    \insertCite{Hay1985,HayBrett1988}{SpawnIndex}, and
    phi is the proportion of spawners that are female.}
  \item{SOK}{List with 3 items:
    nu,
    upsilon, and
    M.}
  \item{surface}{List with 2 items:
    alpha is the regression intercept, and
    beta is the regression slope.}
  \item{macrocystis}{List with 4 items:
    beta,
    gamma,
    delta, and
    epsilon.}
  \item{understory}{List with 4 items:
    alpha,
    beta,
    gamma, and
    delta.}
}}
\usage{
data(pars)
}
\description{
Parameters for Pacific Herring spawn index calculations: converting eggs to
biomass, spawn-on-kelp calculations, as well as surface, Macrocystis, and
understory spawn index calculations.
}
\examples{
data(pars)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CalcEggConversion}} \code{\link{CalcSurfSpawn}}
\code{\link{CalcMacroSpawn}} \code{\link{CalcUnderSpawn}}
}
\keyword{datasets}
