% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pars}
\alias{pars}
\title{Parameters for spawn index calculations.}
\format{
List with items:
\describe{
\item{conversion}{List with items:
\emph{omega} is the number of eggs per kilogram of female spawners
\insertCite{Hay1985,HayBrett1988}{SpawnIndex}, and
\emph{female} is the proportion of spawners that are female.}
\item{sok}{List with items:
\emph{nu} is the proportion of SOK product that is kelp
\insertCite{ShieldsEtal1985}{SpawnIndex},
\emph{upsilon} is the SOK product weight increase during a 24-hour brining
period as a proportion \insertCite{WhyteEnglar1977}{SpawnIndex}, and
\emph{M} is the average weight in kilograms of a fertilized egg
\insertCite{HayMiller1982}{SpawnIndex}.}
\item{surface}{List with items:
\emph{alpha} is the regression intercept
\insertCite{SchweigertEtal1997}{SpawnIndex}, and
\emph{beta} is the regression slope
\insertCite{SchweigertEtal1997}{SpawnIndex}.}
\item{macrocystis}{List with items:
\emph{xi} is the regression slope
\insertCite{HaegeleSchweigert1990}{SpawnIndex},
\emph{gamma} is the regression exponent on egg layers
\insertCite{HaegeleSchweigert1990}{SpawnIndex},
\emph{delta} is the regression exponent on plant height
\insertCite{HaegeleSchweigert1990}{SpawnIndex}, and
\emph{epsilon} is the regression exponent on number of stalks per plant
\insertCite{HaegeleSchweigert1990}{SpawnIndex}.}
\item{understory}{List with items:
\emph{varphi} is the regression slope for substrate
\insertCite{HaegeleEtal1979}{SpawnIndex},
\emph{vartheta} is the regression slope for algae
\insertCite{Schweigert2005}{SpawnIndex},
\emph{varrho} is the regression exponent on number of egg layers
\insertCite{Schweigert2005}{SpawnIndex}, and
\emph{varsigma} is the regression exponent on proportion of algae
\insertCite{Schweigert2005}{SpawnIndex}.}
\item{years}{List with items \insertCite{GrinnellEtalYYYY}{SpawnIndex}:
\emph{survey} is the first year of spawn survey data,
\emph{assess} is the first year of spawn survey data that is reliable for
indexing purposes,
\emph{nine_cats} is the first year that spawn surveyors used nine
(instead of five) spawn intensity categories,
\emph{layers} is the first year that spawn surveyors estimated the number
of egg layers directly (instead of intensity categories), and
\emph{dive} is the first year of the dive survey period.}}
}
\usage{
pars
}
\description{
Parameters for Pacific Herring spawn index calculations: converting eggs to
biomass, spawn-on-kelp (SOK) calculations, as well as surface, Macrocystis,
and understory spawn index calculations. In addition, years in which there
were changes to the survey protocol and data.
}
\examples{
data(pars)
pars
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{load_all_spawn}} \code{\link{calc_egg_conversion}}
\code{\link{calc_biomass_sok}} \code{\link{calc_surf_spawn}}
\code{\link{calc_macro_spawn}} \code{\link{calc_under_spawn}}

Other parameters: 
\code{\link{algae_coefs}},
\code{\link{intensity}},
\code{\link{under_width_facs}}
}
\concept{parameters}
\keyword{datasets}
