% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Cross-walk table for regions and region names.}
\format{
Tibble with rows for regions and columns for attributes.
}
\usage{
regions
}
\description{
Cross-walk table with stock assessment region (SAR) numbers, regions, region
names, and type (i.e., major, minor, special)
\insertCite{GrinnellEtalYYYY}{SpawnIndex}.
}
\note{
JS is not an official SAR; sections 132 and 135 are also in SoG.
Similarly, A10 is not an official SAR; sections 101, 102, and 103 are in
CC.
}
\examples{
data(regions)
regions
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{load_area_data}}

Other parameters: 
\code{\link{algae_coefs}},
\code{\link{intensity}},
\code{\link{pars}},
\code{\link{under_width_facs}}
}
\concept{parameters}
\keyword{datasets}
