% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpawnIndex.R
\docType{package}
\name{SpawnIndex}
\alias{SpawnIndex}
\title{SpawnIndex: Calculating the spawn index for Pacific Herring (\emph{Clupea pallasii}) in
British Columbia, Canada.}
\description{
The SpawnIndex package provides data, parameters, and methods to
calculate the spawn index for Pacific Herring (\emph{Clupea pallasii}) in British
Columbia, Canada. Essentially, spawn index calculations convert spawn survey
observations (e.g., spatial extent, number of egg layers, substrate type) to
the Pacific Herring spawn index. There are three types of spawn survey
observations: surface observations, Macrocystis observations, and understory
observations. In addition, we include methods to convert eggs to biomass,
and estimate spawning biomass in spawn-on-kelp operations. Note that the
'spawn index' is a relative index of spawning biomass.
}
\details{
The SpawnIndex package provides raw data, data, and four families of
functions: load, calculation, check, and utility.
}
\note{
The `spawn index' is a relative index of spawning biomass. Read the
technical report for more details and background on spawn index
calculations \insertCite{GrinnellEtal2022}{SpawnIndex}.
}
\section{Raw data}{
 The raw data is an example database with spawn survey
observations (\code{\link{HerringSpawn}}).
}

\section{Data}{
 The data are region names (\code{\link{regions}}), understory
spawn width correction factors (\code{\link{under_width_facs}}), parameters
for spawn index calculations (\code{\link{pars}}), spawn intensity
categories (\code{\link{intensity}}), algae coefficients
(\code{\link{algae_coefs}}), and a simple feature collection with Section
polygons (\code{\link{sections}}).
}

\section{Load functions}{
 The load functions are \code{\link{load_area_data}},
\code{\link{load_all_spawn}}, \code{\link{load_width}}, and
\code{\link{load_sections}}.
}

\section{Calculation functions}{
 The calculation functions are
\code{\link{eggs_to_sb}}, \code{\link{calc_sok_index}},
\code{\link{dens_surf}}, \code{\link{calc_surf_index}},
\code{\link{eggs_macro}}, \code{\link{calc_macro_index}},
\code{\link{dens_under_sub}}, \code{\link{dens_under_alg}}, and
\code{\link{calc_under_index}}.
}

\section{Check functions}{
 The check functions are
\code{\link{check_numeric}}, \code{\link{check_tibble}}, and
\code{\link{check_where}}.
}

\section{Utility functions}{
 The utility function are
\code{\link{paste_nicely}}, \code{\link{min_na}}, \code{\link{max_na}},
\code{\link{sum_na}}, \code{\link{mean_na}}, \code{\link{wt_mean_na}},
\code{\link{roll_mean_na}}, and \code{\link{unique_na}}.
}

\references{
\insertAllCited{}
}
