% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpawnIndex.R
\docType{package}
\name{SpawnIndex}
\alias{SpawnIndex}
\title{SpawnIndex: A package to calculate the spawn index for Pacific Herring
(\emph{Clupea pallasii}) in British Columbia, Canada.}
\description{
The SpawnIndex package provides data, parameter values, and methods to
calculate the spawn index for Pacific Herring (\emph{Clupea pallasii}) in
British Columbia, Canada. Methods include converting eggs to biomass,
spawn-on-kelp calculations, as well as surface, Macrocystis, and understory
spawn index calculations. Read the technical report for more details and
background on spawn index calculations
\insertCite{GrinnellEtalYYYY}{SpawnIndex}.
}
\details{
The SpawnIndex package provides data, and two categories of functions: load,
and calculations.
}
\note{
The `spawn index' is a relative index of spawning biomass.
}
\section{Data}{
 The data are understory spawn width correction factors
  (\code{\link{under_width_facs}}), parameters for spawn index calculations
  (\code{\link{pars}}), spawn intensity categories (\code{\link{intensity}}),
  and algae coefficients (\code{\link{algae_coefs}}). In addition, there is
  an example database with spawn survey observations
  (\code{\link{HerringSpawn}}).
}

\section{Load functions}{
 The load functions are \code{\link{load_area_data}},
  \code{\link{load_all_spawn}}, and \code{\link{get_width}}.
}

\section{Calculation functions}{
 The calculation functions are
  \code{\link{calc_egg_conversion}}, \code{\link{calc_biomass_sok}},
  \code{\link{calc_surf_spawn}}, \code{\link{calc_macro_spawn}}, and
  \code{\link{calc_under_spawn}}.
}

\references{
\insertAllCited{}
}
