% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpawnIndex.R
\docType{package}
\name{SpawnIndex}
\alias{SpawnIndex}
\title{SpawnIndex: A package to calculate the spawn index for Pacific Herring
(\emph{Clupea pallasii}) in British Columbia, Canada.}
\description{
The SpawnIndex package provides data, parameter values, and methods to
calculate the spawn index for Pacific Herring (\emph{Clupea pallasii}) in
British Columbia, Canada. Methods inluclude converting eggs to biomass,
spawn-on-kelp calculations, as well as surface, Macrocystis, and understory
spawn index calculations. Read the accompanying document for more details on
spawn index calculations \insertCite{GrinnellEtalYYYY}{SpawnIndex}.
}
\note{
The 'spawn index' is a relative index of spawning biomass. In addition,
  32-bit R is required to get data from the 32-bit MS Access database.
}
\references{
\insertAllCited{}
}
