% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_na}
\alias{sum_na}
\title{Calculate the sum.}
\usage{
sum_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{omit_na}{Logical. Omit NAs; default is TRUE.}
}
\value{
A numeric vector.
}
\description{
Calculate the sum; an alternative to \code{sum(x)} that returns \code{NA} if \code{x} is all
\code{NA} or an empty set. Although the sum of an empty is zero, this causes
issues in spawn index calculations.
}
\note{
The sum of an empty set is zero, by definition.
}
\examples{
sum(NA, na.rm = TRUE)
sum_na(NA)
}
\seealso{
\code{\link{sum}}

Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{paste_nicely}()},
\code{\link{unique_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
