% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{underWidthFac}
\alias{underWidthFac}
\title{Understory spawn width correction factors.}
\format{Tibble with 12 rows and 8 columns (one for year, and one for each
  regions).}
\usage{
underWidthFac
}
\description{
Spawn width correction factors for Pacific Herring understory spawn surveys
by region and year. These correction factors account for an error that resulted
in underestimated spawn width for understory dive surveys
\insertCite{ClearyEtal2017,GrinnellEtalYYYY}{SpawnIndex}.
}
\note{
Adjust spawn width for affected understory spawn surveys by multiplying
  the observed width by the corresponding correction fator for the affected
  years.
}
\examples{
data(underWidthFac)
underWidthFac
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{CalcUnderSpawn}}
}
\keyword{datasets}
