% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_na}
\alias{unique_na}
\title{Determine unique values.}
\usage{
unique_na(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
An object of the same type as \code{x}.
}
\description{
Determine unique values; an alternative to \code{unique(x)} that returns \code{NA} if
\code{x} is all \code{NA} or an empty set.
}
\examples{
unique(na.omit(NA))
unique_na(NA)
}
\seealso{
\code{\link{unique}}

Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{paste_nicely}()},
\code{\link{sum_na}()},
\code{\link{wt_mean_na}()}
}
\concept{utility functions}
