% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wt_mean_na}
\alias{wt_mean_na}
\title{Weighted mean
Calculate weighted mean if there are non-NA values, return NA if all values
are NA}
\usage{
wt_mean_na(x, w, omit_na = TRUE)
}
\arguments{
\item{x}{Vector to consider.}

\item{omit_na}{Omit NAs? Logical, default TRUE.}
}
\value{
Numeric.
}
\description{
Weighted mean
Calculate weighted mean if there are non-NA values, return NA if all values
are NA
}
\seealso{
Other utility functions: 
\code{\link{max_na}()},
\code{\link{mean_na}()},
\code{\link{min_na}()},
\code{\link{paste_nicely}()},
\code{\link{roll_mean_na}()},
\code{\link{sum_na}()},
\code{\link{unique_na}()}
}
\concept{utility functions}
